backend/requirements-lite.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pydantic>=2.6.4
email-validator>=2.2.0
tzdata>=2024.2
python-multipart>=0.0.9

# motor (Mongo) is optional and only needed if you supply MONGO_URL and DB_NAME
# Add it with: pip install motor
